requirements.txt
websocket-client==1.6.4
tqdm==4.66.1
colorama==0.4.6
requests==2.31.0 